backend/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
flask-jwt-extended
flask-swagger-ui==4.11.1
flask-sqlalchemy
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.0.8
requests==2.32.3
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.4
pytz==2024.2